Silent debloat softwares: - Move ShutUp10 Download folder - Refactor file to be more safe (removed *) - Put a better description and URLs
README:
- Refactored the common script features

Backup System:
- Rename Restore Point

Opt. Security:
- Renamed tweak output to Security Category

Remove Bloatware Apps:
- Refactor to be more precise when displaying what was removed

Removed .txt files and use 1 instead on /src/tmp/

+++ src/scripts/remove-bloatware-apps.ps1
@@ -138,11 +138,20 @@ function RemoveBloatwareApps() {
     )
 
     ForEach ($Bloat in $Apps) {
-        Write-Host "[-][UWP] Trying to remove $Bloat ..."
-        Get-AppxPackage -AllUsers -Name $Bloat | Remove-AppxPackage # App
-        Get-AppxProvisionedPackage -Online | Where-Object DisplayName -like $Bloat | Remove-AppxProvisionedPackage -Online -AllUsers # Payload
-    }
 
+        If ((Get-AppxPackage -AllUsers -Name $Bloat) -or (Get-AppxProvisionedPackage -Online | Where-Object DisplayName -like $Bloat)) {
+
+            Write-Host "[-][UWP] Trying to remove $Bloat ..."
+            Get-AppxPackage -AllUsers -Name $Bloat | Remove-AppxPackage # App
+            Get-AppxProvisionedPackage -Online | Where-Object DisplayName -like $Bloat | Remove-AppxProvisionedPackage -Online -AllUsers # Payload
+    
+        }
+        Else {
+
+            Write-Warning "[?][UWP] $Bloat was already removed or not found."
+   
+        }
+    }
 }
 
 function Main() {
